Still Buggy But Full Implementation of BlackJack
I completed finishing the implementation of the BlackJackGame (logic) and BlackJackApp (front end), where the game is now playable and partially integrated with game manager. I have also added the required images and music for future implementation. Also, I may need to update the inheritance of methods, as hit and stand are the same for all children of Player. The saving and recording of scores is also unfinished. Lastly, I need to make a game over implementation when the human player goes bankrupt