Major Changes: FileManager, MusicManager, and BlackJack bug fixes (with some additional changes)
Firstly, I consolidated all the file making into a new class FileManager, which takes care of scoring, storing, and hashing (encryption and decryption)

Next, I created a MusicManager that determines which game section is currently being shown, and then plays a unique song for main menu, snake, and blackjack

For Blackjack, I fixed the saving and loading with the FileManager, so that only 1 save string can be made for an account, but multiple saves can be made. I also fixed the bankrupcy for human and computer, where the computer will be skipped and the human will kick you out with the highest value earned. When saving, a password system is implemented to both use to hash/encrypt and decrypt the save file.

Lastly, I added a back button in the login screen, and tried to remove code that is no longer being called.